Moved Notifications to new namespace + pr refactor
